first working version for 10_2_X

+++ JetWidthCalculator/src/JetWidthCalculator.cc
@@ -5,21 +5,21 @@
 
 JetWidthCalculator::JetWidthCalculator(const pat::Jet& jet) {
 
-  double pfCand1pt_      = 0;
-  double pfCand2pt_      = 0;  
+  float pfCand1pt_      = 0.;
+  float pfCand2pt_      = 0.;  
  
-  double etSum_          = 0;
-  double etaSum_         = 0;
-  double etaSqSum_       = 0;
-  double phiSum_         = 0;
-  double phiSqSum_       = 0;
+  float etSum_          = 0.;
+  float etaSum_         = 0.;
+  float etaSqSum_       = 0.;
+  float phiSum_         = 0.;
+  float phiSqSum_       = 0.;
   
-  etaWidth_              = 0;
-  phiWidth_              = 0;
+  etaWidth_              = 0.;
+  phiWidth_              = 0.;
 
-  ptSum_                 = 0;
-  pfCand12PtSum_         = 0;
-  pt12ratio_             = 0;
+  ptSum_                 = 0.;
+  pfCand12PtSum_         = 0.;
+  pt12ratio_             = 0.;
 
   nPhotons_              = 0;
   nCHPions_              = 0;
@@ -32,7 +32,7 @@ JetWidthCalculator::JetWidthCalculator(const pat::Jet& jet) {
   ConstPdgId_      .clear();
   MiscPdgId_       .clear();
 
-  std::vector<std::pair<double, const reco::Candidate*>> jetDaughters;
+  std::vector<std::pair<float, const reco::Candidate*>> jetDaughters;
   const uint32_t packedCands = jet.numberOfDaughters();
   
   for(uint32_t i = 0; i < packedCands; i++) {
@@ -85,12 +85,12 @@ JetWidthCalculator::JetWidthCalculator(const pat::Jet& jet) {
   pt12ratio_     = (pfCand12PtSum_/ptSum_);
   
   if(etSum_ < 0.000001) etSum_ = 0.000001; // To avoid NaNs
-  double etaAve_   = etaSum_/etSum_;
-  double etaSqAve_ = etaSqSum_/etSum_;
+  float etaAve_   = etaSum_/etSum_;
+  float etaSqAve_ = etaSqSum_/etSum_;
   etaWidth_        = sqrt(etaSqAve_ - (etaAve_ * etaAve_));
   
-  double phiAve_   = phiSum_ / etSum_;
-  double phiSqAve_ = phiSqSum_ / etSum_;
+  float phiAve_   = phiSum_ / etSum_;
+  float phiSqAve_ = phiSqSum_ / etSum_;
   phiWidth_        = sqrt(phiSqAve_ - (phiAve_ * phiAve_));
   
 }
